requirements.txt
paho-mqtt
python-miio
